+++ app/build.gradle
@@ -23,7 +23,9 @@ android {
 }
 
 dependencies {
+    implementation project(':repository-stub')
     implementation project(':repository-api')
+
     //lombok
     compileOnly 'org.projectlombok:lombok:1.18.2'
     annotationProcessor 'org.projectlombok:lombok:1.18.2'
@@ -31,6 +33,9 @@ dependencies {
     testAnnotationProcessor 'org.projectlombok:lombok:1.18.2'
     androidTestImplementation 'org.projectlombok:lombok:1.18.2'
     androidTestAnnotationProcessor 'org.projectlombok:lombok:1.18.2'
+
+    implementation 'org.slf4j:slf4j-api:1.7.25'
+    implementation 'com.github.tony19:logback-android:1.1.1-12'
 
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support.constraint:constraint-layout:1.1.2'
