+++ app/src/main/java/com/example/pto6/ofc/service/DBHelperStub.java
@@ -3,6 +3,7 @@ package com.example.pto6.ofc.service;
 import com.example.pto6.ofc.model.Credit;
 import com.example.pto6.ofc.model.Debit;
 import com.example.repository.Repository;
+import com.example.repository.stub.RepositoryStub;
 
 import java.util.ArrayList;
 import java.util.List;
