+++ repository-api/build.gradle
@@ -1,7 +1,5 @@
-apply plugin: 'java'
+apply plugin: 'java-library'
 
 dependencies {
-    implementation 'com.google.code.gson:gson:2.8.5'
-
-    testImplementation 'junit:junit:4.12'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
 }
