+++ repository-stub/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'io.franzbecker.gradle-lombok' version '1.14'
+    id 'java-library'
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation project(':repository-api')
+
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
+}
+
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
